requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dictdiffer==0.9.0
exceptiongroup==1.2.0
fastapi==0.110.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
idna==3.6
iso8601==1.1.0
numpy==1.26.4
pandas==2.2.1
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tomlkit==0.12.4
tortoise-orm==0.20.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.29.0
